Revert to Minion
This document was originally set in Minion. Rather late in the
process, the university introduced a new set of rules governing
the format of a PhD thesis that, bizarrely, mandated the use of
sans serif type. I believe this rule was eventually overturned
following the protests of faculty in the sort of mathematical
disciplines that have never developed the habit of using sans-
serif, but not before I had reformatted the thesis in Myriad.

However I find it a more attractive document in its original
setting, so this seems a reasonable opportunity to revert.