Add PDF files to repository
It would be nice to have the final PDF here, for the benefit
of anyone who does not own the necessary fonts.

In the process of doing this, I realised that the PDF diagrams
from the semicc project had been omitted incorrectly, since
they were caught by the blanket “*.pdf” that I had in .gitignore.